Move Status enum into its own folder for organization purposes
