+++ backend/src/main/java/com/codify/backend/dto/JobApplicationRequest.java
@@ -1,8 +1,7 @@
 package com.codify.backend.dto;
 
+import com.codify.backend.enums.Status;
 import java.time.LocalDate;
-
-import com.codify.backend.Status;
 
 public record JobApplicationRequest(
 	String positionTitle,
